WIP: Detecting text positions

+++ src/pdf2reader/data_structures.py
@@ -3,9 +3,9 @@ from dataclasses import dataclass
 
 @dataclass
 class Box:
-    x0: int
-    y0: int
-    x1: int
-    y1: int
+    x0: float
+    y0: float
+    x1: float
+    y1: float
     color: str
     on_click: callable
